Load all scenes of a gltf file, not only the main one
The adding of geometries to the UI menu has been moved to the callback so that it becomes possible to create several geometries from a given file, one per scene.
The name provided is now disregarded if it's empty, and the names of the scenes in the geometry are used in such a case.
They may contain a path with the ' > ' syntax, these path will be respected to build a hierarchy in the menu